pyproject.toml
[tool.poetry]
name = "ringmaster.show"
version = "0.2.4+c2db814"
description = "The world is a circus and you are the ringmaster!"
authors = ["Geoff Williams <geoff@declarativesystems.com>"]
license = "Apache 2.0"
readme = "README.md"
homepage = "https://ringmaster.show"
repository = "https://github.com/declarativesystems/ringmaster"
documentation = "https://ringmaster.show/"
packages = [
    { include = "ringmaster" },
]
include = ["ringmaster/version.py"]

[tool.poetry.dependencies]
python = "^3.8"
python-dateutil = "^2.8.1"
boto3 = "^1.17.24"
loguru = "^0.5.3"
docopt = "^0.6.2"
halo = "^0.0.31"
snowflake-connector-python = "^2.4.1"
snakecase = "^1.0.1"
cfn-flip = "^1.2.3"
Jinja2 = "^2.11.3"
python-cloudflare = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ringmaster = "ringmaster.cli:main"